Add context support
