+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - `orm.URLValues` is renamed to `orm.URLFilters`. It is no longer adds ORDER clause.
  - `orm.Pager` is renamed to `orm.Pagination`.
  - Support for net.IP and net.IPNet.
+ - Support for context.Context.
 
 ## v5
